Final set of bugfixes; All tests now pass.
- Default internal csv loads to guessing types
- R.csv now predicts as Int rather than short

+++ vizier/backend/test/src/info/vizierdb/commands/PythonSaveDatasetSpec.scala
@@ -51,7 +51,7 @@ class PythonSaveDatasetSpec
       val df = project.dataframe("R")
       val data = df.collect.toSeq
       println(s"Read artifact #${project.artifact("R").id}")
-      data.map { _.getShort(1) } must contain(42)
+      data.map { _.getInt(1) } must contain(42)
       data.filterNot { _.isNullAt(1) } must haveSize(data.size)
       df.schema(3).name must beEqualTo("D")
       df.schema(3).dataType must beEqualTo(FloatType)
